support for [country]/[map]
